chore: Start font manifest support on iOS
